Implement pagination

+++ iamjob/src/main/java/nick/iamjob/data/PositionsRepository.kt
@@ -1,11 +1,16 @@
 package nick.iamjob.data
 
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import io.reactivex.Completable
 import nick.core.di.ApplicationScope
+import nick.core.util.Event
 import nick.data.dao.PositionsDao
 import nick.data.model.Position
 import nick.data.model.Search
+import nick.iamjob.util.PositionsLoadingState
 import nick.networking.GitHubJobsService
+import timber.log.Timber
 import javax.inject.Inject
 
 @ApplicationScope
@@ -14,17 +19,31 @@ class PositionsRepository @Inject constructor(
     private val positionsDao: PositionsDao
 ) {
 
-    fun search(search: Search): Completable = with(search) {
+    private val _exhausedPages = MutableLiveData<Event<Unit>>()
+    val exhaustedPages: LiveData<Event<Unit>> get() = _exhausedPages
+
+    fun search(search: Search, loadingState: PositionsLoadingState): Completable = with(search) {
         service.fetchPositions(
             description = description.ifEmpty { null },
             location = location.ifEmpty { null },
             isFullTime = isFullTime,
             page = page
         ).flatMapCompletable { fetchedPositions ->
-            // Mark all cached positions as stale -- we don't want them showing up in search results
-            // if they're not part of the remotely fetched result set
+
+            if (fetchedPositions.isEmpty()) {
+                Timber.d("Exhausted pages for: $search")
+                _exhausedPages.postValue(Event(Unit))
+                return@flatMapCompletable Completable.complete()
+            }
+
             val cachedPositions = positionsDao.queryCachedBlocking().map {
-                it.copy(isFresh = false)
+                if (loadingState is PositionsLoadingState.Paginating) {
+                    it
+                } else {
+                    // Mark all cached positions as stale -- we don't want them showing up in search results
+                    // if they're not part of the remotely fetched result set
+                    it.copy(isFresh = false)
+                }
             }.toMutableList()
 
             // Apply cached position states to the newly fetched positions
@@ -44,10 +63,17 @@ class PositionsRepository @Inject constructor(
                 it.addAll(cachedPositions)
             }
 
-            // Insert will replace (effectively updating) any positions with the same ID
-            deleteAllNonCachableThenInsert(reconciledPositions)
+            if (loadingState is PositionsLoadingState.Paginating) {
+                insert(reconciledPositions)
+            } else {
+                // Insert will replace (effectively updating) any positions with the same ID
+                deleteAllNonCachableThenInsert(reconciledPositions)
+            }
         }
     }
+
+    private fun insert(positions: List<Position>): Completable =
+        Completable.fromAction { positionsDao.insert(positions) }
 
     private fun deleteAllNonCachableThenInsert(positions: List<Position>): Completable =
         Completable.fromAction { positionsDao.deleteAllNonCachableThenInsert(positions) }
